requirements.txt
imaplib
email
PyYAML
beautifulsoup4
unidecode
